Extract copy handler and move to existing div
